# Improved performance of sum reduction