--- a/tinygrad/codegen/cstyle.py
+++ b/tinygrad/codegen/cstyle.py
@@ -20,6 +20,7 @@ class CStyleLanguage(NamedTuple):
   barrier: str = ""
   atomic_prefix: str = ""
   atomic_add: str = ""
+  simd_sum: str = ""
   gid: List[str] = []
   lid: List[str] = []
   extra_args: List[str] = []
@@ -159,6 +160,15 @@ def uops_to_cstyle(uops:List[UOp], lang:CStyleLanguage) -> Tuple[str, List[int],
       kk(f"c.thread_elements()[0] = {vin[4].render()}; c.thread_elements()[1] = {vin[5].render()};")
       kk("simdgroup_multiply_accumulate(c, a, b, c);")
       kk(f"{vin[4].render()} = c.thread_elements()[0]; {vin[5].render()} = c.thread_elements()[1]; }}")
+    elif uop == UOps.LOCAL_REDUCE:
+      local_index = Variable.sum(vin[0]).render(render_cl)
+      kk(f"{{ int lane = {local_index} % warp_size;")
+      kk(f"int wid = {local_index} / warp_size;")
+      kk(f"{lang.simd_sum.format(vin[1].render())};")
+      kk(f"if (lane == 0) {vin[2].name}[wid]={vin[1].render()};")
+      kk(f"{lang.barrier};")
+      kk(f"{vin[1].render()} = ({local_index} < {vin[3]} / warp_size) ? {vin[2].name}[lane] : 0;")
+      kk(f"if (wid == 0) {lang.simd_sum.format(vin[1].render())}; }}")
     elif uop == UOps.ALU:
       assert newvar is not None
       kk(f"{lang.generic_var_prefix if newvar not in vin else ''}{newvar.render(newvar not in vin and lang.generic_var_prefix == '')} = {lang.code_for_op[args](*[x.render() for x in vin])};")
@@ -198,6 +208,7 @@ class CStyleCodegen(Linearizer):
   supports_float4: bool = True
   supports_float4_alu: bool = True
   supports_atomics: bool = False
+  supports_fast_local_reduce: bool = False
 
   # for renaming
   kernel_cnt: Final[DefaultDict[str, int]] = collections.defaultdict(int)