… .github/workflows/test.yml
… test/test_ops.py
… test/test_speed_v_torch.py
… tinygrad/codegen/wgsl.py
… tinygrad/helpers.py
… tinygrad/lazy.py
… tinygrad/ops.py
… tinygrad/runtime/lib.py
… tinygrad/runtime/ops_cuda.py
… tinygrad/runtime/ops_hip.py
… tinygrad/runtime/ops_metal.py