This merge request introduces support for atomics and fast late-reduce operations. These enhancements enable us to achieve performance levels closer to, and in some cases, even surpassing those of PyTorch CUDA. Previously, we were considerably slower than PyTorch on CUDA by approximately 15x, but with the introduction of a two-kernel approach for the reduction process, we have reduced the gap to just 2x-3x slower than CUDA. The inclusion of atomics and fast reduce now allows us to outperform PyTorch or come very close to its speed.

2 kernels + atomics + fast late-reduce:
```
 sum                             1024x 1024    0.11 ms (    9.97 GFLOPS    41.07 GB/s) in torch,    0.07 ms (   14.89 GFLOPS    61.38 GB/s) in tinygrad,    0.67x faster       1.08 MOPS     4.46 MB
 sum                             2048x 2048    0.11 ms (   40.73 GFLOPS   167.85 GB/s) in torch,    0.10 ms (   42.95 GFLOPS   177.02 GB/s) in tinygrad,    0.95x faster       4.33 MOPS    17.83 MB
 sum                             4096x 4096    0.23 ms (   74.64 GFLOPS   307.61 GB/s) in torch,    0.20 ms (   85.57 GFLOPS   352.64 GB/s) in tinygrad,    0.87x faster      17.30 MOPS    71.30 MB
 sum                             8192x 8192    0.66 ms (  104.19 GFLOPS   429.41 GB/s) in torch,    0.70 ms (   98.56 GFLOPS   406.17 GB/s) in tinygrad,    1.06x slower      69.21 MOPS   285.21 MB
```
Just 2 kernels:
```
 sum                             1024x 1024    0.11 ms (    9.72 GFLOPS    39.03 GB/s) in torch,    0.06 ms (   17.60 GFLOPS    70.69 GB/s) in tinygrad,    0.55x faster       1.05 MOPS     4.23 MB
 sum                             2048x 2048    0.11 ms (   39.95 GFLOPS   160.42 GB/s) in torch,    0.09 ms (   45.47 GFLOPS   182.59 GB/s) in tinygrad,    0.88x faster       4.21 MOPS    16.91 MB
 sum                             4096x 4096    0.25 ms (   67.18 GFLOPS   269.78 GB/s) in torch,    0.49 ms (   34.40 GFLOPS   138.14 GB/s) in tinygrad,    1.95x slower      16.84 MOPS    67.63 MB
 sum                             8192x 8192    0.66 ms (  101.45 GFLOPS   407.37 GB/s) in torch,    2.19 ms (   30.81 GFLOPS   123.72 GB/s) in tinygrad,    3.29x slower      67.37 MOPS   270.53 MB
```

Atomics have been successfully implemented for CUDA, HIP, and Metal backends. These optimizations have proven to enhance performance on larger GPUs, as they enable us to move part of the reduction process to global dimensions, resulting in better GPU utilization. As for fast late-reduce this is what torch uses in their CUDA kernel (nvidia also recommends simd shuffles in [their article](https://developer.nvidia.com/blog/faster-parallel-reductions-kepler/)). It seems to be possible to implement for HIP backend (they have the same instruction), but I just put a TODO, since I cannot test this for now.

Another part in this series is device info which should allow linearlizer to generate dims based on device specs. This implementation is expected to be highly beneficial, as it will enable tinygrad to cater to a wide range of devices, from the M1 to the RTX4090. I am not sure about the way I implemented this, maybe you have any suggestions on this.

This merge request has a dependency on #1187. Currently, I have cherry-picked the commit, but once it is merged into the master branch, I will be able to rebase.