@tinyb0t 2023-07-19
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/codegen/cstyle.py                25               6
tinygrad/codegen/linearizer.py            64              12
tinygrad/codegen/wgsl.py                   4               4
tinygrad/helpers.py                        5               0
tinygrad/lazy.py                          15               1
tinygrad/ops.py                            3               3
tinygrad/runtime/lib.py                    9               1
tinygrad/runtime/ops_cuda.py              12               3
tinygrad/runtime/ops_hip.py                3               0
tinygrad/runtime/ops_metal.py             14               3
------------------------------------------------------------
total                                    154              33
------------------------------------------------------------
lines added in the tinygrad folder: 121
```