@nimlgen 2023-07-19
P.S. These changes made me to disable some more metal tests for CI. I spent some time on this and it seems that the problem occurs when there are a lot of live registers inside one kernel (when do heavy upcasting). It is possible to limit this, but I don't know if it worth adding complexity just for CI and at the same time nobody knows the truth of the failures.