@g1y5x3 2023-07-19
> Another part in this series is device info which should allow linearlizer to generate dims based on device specs. This implementation is expected to be highly beneficial, as it will enable tinygrad to cater to a wide range of devices, from the M1 to the RTX4090. I am not sure about the way I implemented this, maybe you have any suggestions on this.

Been working on fixing cuda dude to dim limits on `examples/hlb_cifar10.py` #1224 and I was thinking one of the better solution might be to approach from `linearizer` since there are many kernels were trying to allocate on `global.y` and `global.z` which have significantly smaller sizes compare to `global.x`

I haven't read into your code yet but I cloned the repo and tried on the branch. The test still fails as of now `GPU=1 CUDA=1 BS=512 STEPS=1000 python examples/hlb_cifar10.py` on a 3090 with the same problem.