@nimlgen 2023-07-19
@g1y5x3, I believe it would be beneficial to limit dimension sizes based on device info. With this enhancement, we should be able to optimize by pulling some computations into local dimensions from the global ones. However, we should also consider the scenario where we exceed both global and local size limitations. In such cases, it might be necessary to add an inner kernel loop to handle these situations efficiently. This would allow us to not fail, since we could always unload global+local to inner loop. But in this particular case, setting local dimensions up to 512 (maybe 1024?) seems to be the most suitable option for optimal performance.