@g1y5x3 2023-07-19
I was trying to limit it by checking only in `ops_cuda.py` instead of creating global+local inner loop in #1235. I was just splitting up the same kernels with added offsets onto the corresponding dim. For some reason it is failing when `BS=1024`.

But I think we can separate out the device info part of this PR and added in the global+local inner loop and test with dim limit to see which results in better performance.